+++ src/main/java/com/hamdihawari/server/project/imageGroup/service/ImageTranslationService.java
@@ -2,20 +2,23 @@ package com.hamdihawari.server.project.imageGroup.service;
 
 import com.hamdihawari.server.project.imageGroup.dto.ImageTranslationDTO;
 import com.hamdihawari.server.project.imageGroup.entity.ImageTranslation;
-
 import com.hamdihawari.server.project.imageGroup.repository.ImageTranslationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class ImageTranslationService {
 
+    private final ImageTranslationRepository imageTranslationRepository;
+
     @Autowired
-    private ImageTranslationRepository imageTranslationRepository;
+    public ImageTranslationService(ImageTranslationRepository imageTranslationRepository) {
+        this.imageTranslationRepository = imageTranslationRepository;
+    }
 
     public List<ImageTranslationDTO> getAllTranslations() {
         return imageTranslationRepository.findAll().stream()
@@ -23,10 +26,8 @@ public class ImageTranslationService {
                 .collect(Collectors.toList());
     }
 
-    public ImageTranslationDTO getTranslationById(Long id) {
-        return imageTranslationRepository.findById(id)
-                .map(this::convertToDTO)
-                .orElse(null);
+    public Optional<ImageTranslationDTO> getTranslationById(Long id) {
+        return imageTranslationRepository.findById(id).map(this::convertToDTO);
     }
 
     public ImageTranslationDTO createTranslation(ImageTranslationDTO translationDTO) {
@@ -35,23 +36,27 @@ public class ImageTranslationService {
         translation.setData(translationDTO.getData());
         translation.setImageSubject(translationDTO.getImageSubject());
         translation.setImageDescription(translationDTO.getImageDescription());
-        translation.setImageResource(translationDTO.getImageResource());
+        /*translation.setImageResource(translationDTO.getImageResource());*/
         return convertToDTO(imageTranslationRepository.save(translation));
     }
 
-    public ImageTranslationDTO updateTranslation(Long id, ImageTranslationDTO translationDTO) {
+    public Optional<ImageTranslationDTO> updateTranslation(Long id, ImageTranslationDTO translationDTO) {
         return imageTranslationRepository.findById(id).map(existingTranslation -> {
             existingTranslation.setLanguage(translationDTO.getLanguage());
             existingTranslation.setData(translationDTO.getData());
             existingTranslation.setImageSubject(translationDTO.getImageSubject());
             existingTranslation.setImageDescription(translationDTO.getImageDescription());
-            existingTranslation.setImageResource(translationDTO.getImageResource());
+            /*existingTranslation.setImageResource(translationDTO.getImageResource());*/
             return convertToDTO(imageTranslationRepository.save(existingTranslation));
-        }).orElse(null);
+        });
     }
 
-    public void deleteTranslation(Long id) {
-        imageTranslationRepository.deleteById(id);
+    public boolean deleteTranslation(Long id) {
+        if (imageTranslationRepository.existsById(id)) {
+            imageTranslationRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     private ImageTranslationDTO convertToDTO(ImageTranslation translation) {
@@ -61,7 +66,7 @@ public class ImageTranslationService {
         dto.setData(translation.getData());
         dto.setImageSubject(translation.getImageSubject());
         dto.setImageDescription(translation.getImageDescription());
-        dto.setImageResource(translation.getImageResource());
+        /*dto.setImageResource(translation.getImageResource());*/
         return dto;
     }
 }
